move to /src/cli
