+++ src/cli/internal/cmd/lifecycle/vmop/vmop.go
